Fix file path in MakeVideoThumbnail.php and add conditional migration rollback

+++ database/migrations/2024_05_18_143513_add_loser_id_into_user_game_results.php
@@ -27,6 +27,9 @@ class AddLoserIdIntoUserGameResults extends Migration
      */
     public function down()
     {
+        if(app()->environment('testing')){
+            return;
+        }
         Schema::table('user_game_results', function (Blueprint $table) {
             $table->dropForeign(['loser_id']);
             $table->dropColumn(['loser_id', 'loser_name', 'candidates']);
